# backend/src/utils/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1.0",
    "title": "demoApplication"
  },
  "host": "demo.test.com",
  "schemes": ["https", "http"],
...
  "paths": {
...
    "/api/transactions/{offset}": {
      "get": {
        "tags": ["Transactions"],
        "summary": "Returns user's transactions",
        "description": "",
        "operationId": "getTransactions",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "offset",
            "in": "path",
            "description": "not required!",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/getTransactions"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/transactions/sender": {
      "get": {
        "tags": ["Transactions"],
        "summary": "Returns user's as sender transactions",
        "description": "",
        "operationId": "getSTransactions",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/getSenderTransactions"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/transactions/recipient": {
      "get": {
        "tags": ["Transactions"],
        "summary": "Returns user's as recipient transactions",
        "description": "",
        "operationId": "getRTransactions",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/getRecipientTransactions"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/transactions/{id}/key": {
      "get": {
        "tags": ["Transactions"],
        "summary": "Returns user's as recipient transactions",
        "description": "",
        "operationId": "key",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "schema": {
              "$ref": "#/definitions/getTransactionKey"
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic"
    },
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "getTransactions": {
      "type": "object",
      "properties": {
        "transactions": {
          "type": "array",
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "createdDate": {
                  "type": "string",
                  "format": "date-time"
                },
                "amountMoney": {
                  "type": "string"
                },
                "transferTitle": {
                  "type": "string"
                },
                "currency": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "format": "int64"
                    },
                    "name": {
                      "type": "string"
                    }
                  }
                },
                "recipient": {
                  "type": "object",
                  "properties": {
                    "accountBill": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "format": "int64"
                        },
                        "name": {
                          "type": "string"
                        },
                        "surname": {
                          "type": "string"
                        }
                      }
                    }
                  }
                },
                "sender": {
                  "type": "object",
                  "properties": {
                    "accountBill": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "format": "int64"
                        },
                        "name": {
                          "type": "string"
                        },
                        "surname": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "count": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "getSenderTransactions": {
      "type": "object",
      "properties": {
        "senderTransactions": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "transaction_createdDate": {
                "type": "string",
                "format": "date-time"
              },
              "transaction_amountMoney": {
                "type": "string"
              },
              "transaction_transferTitle": {
                "type": "string"
              },
              "currency_name": {
                "type": "string"
              },
              "user_name": {
                "type": "string"
              },
              "user_surname": {
                "type": "string"
              },
              "user_id": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        }
      }
    },
    "getRecipientTransactions": {
      "type": "object",
      "properties": {
        "recipientTransactions": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "transaction_createdDate": {
                "type": "string",
                "format": "date-time"
              },
              "transaction_amountMoney": {
                "type": "string"
              },
              "transaction_transferTitle": {
                "type": "string"
              },
              "currency_name": {
                "type": "string"
              },
              "user_name": {
                "type": "string"
              },
              "user_surname": {
                "type": "string"
              },
              "user_id": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        }
      }
    },
    "getTransactionKey": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "authorizationKey": {
          "type": "string"
        }
      }
    },
...
  }
}